[Back Key Handling] Added support for handling back key on Fragments inside a ViewPager.
Little details: 
Back key will be obtained to the activity only. It needs to be delegated to the respective fragments. This is done by using a custom function in Fragments - handleBackKey, achieved by extending the Fragment class. 
Posted a reply to a stackOverFlow thread: See my response in link below for details about current implementation:

http://stackoverflow.com/questions/13185476/how-to-handle-back-button-using-view-pager/38588453#38588453

Also referred to answer by FraZer on link: http://stackoverflow.com/questions/18609261/getting-the-current-fragment-instance-in-the-viewpager. This is to solve the issue of FragmentPagerAdapter returning new instance of Fragment every time it gets a call on getItem(int index)