[Refactoring][Edit call log] Refactored the rendering code to multiple methods for improved readability.
Also made modification in contact image setting logic. Sometimes, if the contact is in the address book and if there is no picture associated with it, the image was not getting set. Now this is addressed by checking the image on the ImageView after setting the url. If there is no image set already, default image is set.
Other logic remains the same