[Fake call log][OPEN SOURCE CODE ADDED] Implemented feature of updating the contact details (image and name) from the number entered by user. Also added ImageCache to be used for list (later)
ImageCache is also added. Using the code available in android sample application. This is open source and the credit needs to be provided on app release. Those classes have their header maintained as required (see ImageCache.java and ImageLoader.java for example).
The code is taken from a sample project available in link :
https://developer.android.com/training/contacts-provider/index.html?hl=ko
(Android official guide for “Accessing Contacts Data”)